+++ src/routes/+page.server.ts
@@ -1,4 +1,4 @@
-import type { Actions, PageServerLoad } from './$types';
+import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = () => {
 	return {
@@ -10,35 +10,3 @@ export const load: PageServerLoad = () => {
 		}
 	};
 };
-
-export const actions: Actions = {
-	default: async ({ cookies, request }) => {
-		const data = await request.formData();
-		const name = data.get('name');
-		const rating = data.get('rating');
-		const message = data.get('message');
-
-		console.log('did i get in');
-		if (cookies.get('ratingSent') == 'true') {
-			console.log('Message already sent');
-			cookies.delete('ratingSent');
-			console.log('cookie destroyed');
-			return { success: false };
-		} else {
-			console.log(name);
-			console.log(rating);
-			console.log(message);
-			// Set the cookie sent to prevent spamming
-			console.log('why this dont work');
-			// cookies.set('ratingSent', 'true', {
-			// 	path: '/',
-			// 	httpOnly: true,
-			// 	sameSite: 'strict',
-			// 	maxAge: 60 * 60 * 24
-			// });
-			console.log('is this true');
-			return { success: true };
-		}
-		return { succes: true };
-	}
-} satisfies Actions;
