+++ src/routes/formSubmit/+page.svelte
@@ -1,4 +1,11 @@
-<script>
-	import { redirect } from '@sveltejs/kit';
-	throw redirect(303, '/');
+<script lang="ts">
+	import type { ActionData } from '../$types';
+	import { goto } from '$app/navigation';
+	import { browser } from '$app/environment';
+
+	export let form: ActionData;
+	console.log(form);
+	if (browser) {
+		goto('/');
+	}
 </script>
